+++ .vscode/targets.log
@@ -1,4 +1,6 @@
 make all --print-data-base --no-builtin-variables --no-builtin-rules --question
+make: *** No rule to make target 'all'.  Stop.
+
 # GNU Make 4.2.1
 # Built for x86_64-pc-linux-gnu
 # Copyright (C) 1988-2016 Free Software Foundation, Inc.
@@ -6,7 +8,7 @@ make all --print-data-base --no-builtin-variables --no-builtin-rules --question
 # This is free software: you are free to change and redistribute it.
 # There is NO WARRANTY, to the extent permitted by law.
 
-# Make data base, printed on Thu Aug 25 21:57:21 2022
+# Make data base, printed on Fri Aug 26 23:10:32 2022
 
 # Variables
 
@@ -71,7 +73,7 @@ PATH = /usr/local/bin:/usr/bin:/bin:/usr/local/games:/usr/games
 # environment
 GPG_AGENT_INFO = /run/user/1000/gnupg/S.gpg-agent:0:1
 # environment
-SESSION_MANAGER = local/snd1:@/tmp/.ICE-unix/1556,unix/snd1:/tmp/.ICE-unix/1556
+SESSION_MANAGER = local/snd1:@/tmp/.ICE-unix/1516,unix/snd1:/tmp/.ICE-unix/1516
 # environment
 XDG_RUNTIME_DIR = /run/user/1000
 # environment
@@ -87,7 +89,7 @@ XDG_SESSION_DESKTOP = lightdm-xsession
 # environment
 SSH_AUTH_SOCK = /run/user/1000/keyring/ssh
 # environment
-GIO_LAUNCHED_DESKTOP_FILE_PID = 5216
+GIO_LAUNCHED_DESKTOP_FILE_PID = 5853
 # automatic
 %F = $(notdir $%)
 # environment
@@ -101,7 +103,7 @@ PWD = /home/sander/Bureaublad/SanderOS64
 # environment
 XDG_SEAT = seat0
 # environment
-SSH_AGENT_PID = 1604
+SSH_AGENT_PID = 1564
 # environment
 ORIGINAL_XDG_CURRENT_DESKTOP = MATE
 # environment
@@ -201,7 +203,7 @@ LANG = C
 # environment
 SANDEROS_MOUNT_POINT = /mnt/5DDC-4569
 # environment
-VSCODE_PID = 5216
+VSCODE_PID = 5853
 # variable set hash-table stats:
 # Load=96/1024=9%, Rehash=0, Collisions=8/118=7%
 
@@ -269,8 +271,6 @@ GNUmakefile:
 # strcache performance: lookups = 26 / hit rate = 11%
 # hash-table stats:
 # Load=23/8192=0%, Rehash=0, Collisions=1/26=4%
-# Finished Make data base on Thu Aug 25 21:57:21 2022
+# Finished Make data base on Fri Aug 26 23:10:32 2022
 
  
-make: *** No rule to make target 'all'.  Stop.
-
