+++ drivers/build_drivers
@@ -1,3 +1,4 @@
 $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -mgeneral-regs-only -c ps_2_keyboard.c -o ../bin/sanderos/drivers/ps2key.sys
 $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -mgeneral-regs-only -c xhci.c -o ../bin/sanderos/drivers/xhci.sys
 $SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -mgeneral-regs-only -c test.c -o ../bin/sanderos/drivers/test.sys
+$SANDEROS_CROSS_LOCATION/x86_64-elf-gcc -ffreestanding -mcmodel=large -mno-red-zone -mno-mmx -mno-sse -mno-sse2 -mgeneral-regs-only -c rtl.c -o ../bin/sanderos/drivers/rtl.sys
